Some changes

+++ src/uk/ac/nott/mrl/homework/server/SetName.java
@@ -44,10 +44,10 @@ public class SetName extends HttpServlet
 		rpc.disconnect();
 		
 		final String sinceString = request.getParameter("since");
-		double since = 0;
+		long since = 0;
 		try
 		{
-			since = Double.parseDouble(sinceString);
+			since = Long.parseLong(sinceString);
 		}
 		catch (final Exception e)
 		{
